chore: move to components folder, add nossr
